+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'osbase'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
+name             'mycookbook'
+maintainer       'sampleb3'
+maintainer_email ''
 license          'All rights reserved'
-description      'Installs/Configures osbase'
+description      'Installs/Configures mycookbook'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'