+++ recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: osbase
+# Cookbook Name:: mycookbook
 # Recipe:: default
 #
 # Copyright 2014, YOUR_COMPANY_NAME