+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-osbase CHANGELOG
+mycookbook CHANGELOG
 ================
 
-This file is used to list changes made in each version of the osbase cookbook.
+This file is used to list changes made in each version of the mycookbook cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of osbase
+- [sampleb3] - Initial release of mycookbook
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.